00_GCP_Projects/06_GCP_Document_OCR/document_ocr/pyproject.toml
[tool.poetry]
name = "document-ocr"
version = "0.1.0"
description = ""
authors = ["JaviPSanchez <javier.palomino-sanchez@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-documentai = "^3.0.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
